Workaround for slow status rendering.
Not sure what's the reason, but attempt to render statusbar with many
colorcodes slows down dwm a lot. In dependent of the tool which writes
the string into the root window. Dwm is also slow in case you write root
window direct with some color codes like:

    xsetroot -name "$(echo "1\x012\x023\x033\x044\x011\x022\x033\x044\x055\x011\x022\x033\x044\x055\x011\x022\x033\x044\x055\x01foo")"

If the tags are toggeled and the status bar is redrawn this took a lot
of time to get the real text width without the color codes.

This patch changes the way how the width is calculated. Instead of
running drw_text (is slow for some unknown reason) and subtracting x
times the width of the found colorcodes, the given text is copied into a
new variable without the color codes and given to drw_text (fast).